Added a post-search parameter optimization step when propQuartets<1 (#22)
Added a post-search parameter optimization step when propQuartets<1 so that network likelihood scores are comparable even when a subset of quartets is used to improve search speeds. This makes the log likelihoods of networks inferred with propQuartets<1 identical if and only if the topology discovered in the search is the same between networks. Tests have been updated to reflect these new log likelihood values, and an additional message in the log has been added that mentions the post-search optimization.

---------

Co-authored-by: jjustison <justisnojo@gmail.com>
Co-authored-by: Cecile Ane <cecileane@users.noreply.github.com>
Co-authored-by: Claudia Solis-Lemus <solislemus@wisc.edu>